fix: refactoring page component folders
