+++ src/app/modules/address/address.routing.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { AddressHomeComponent } from './page/address-home/address-home.component';
+import { AddressHomeComponent } from './pages/address-home/address-home.component';
 
 export const ADDRESS_ROUTES: Routes = [
   {
