+++ src/app/modules/customer-vehicle/customer-vehicle.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CustomerVehicleTableComponent } from './components/customer-vehicle-table/customer-vehicle-table.component';
-import { CustomerVehicleHomeComponent } from './page/customer-vehicle-home/customer-vehicle-home.component';
+import { CustomerVehicleHomeComponent } from './pages/customer-vehicle-home/customer-vehicle-home.component';
 import { CUSTOMER_VEHICLE_ROUTES } from './customer-vehicle.routing';
 import { RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
