+++ src/app/modules/customer-vehicle/customer-vehicle.routing.ts
@@ -1,5 +1,5 @@
 import { Routes, RouterModule } from '@angular/router';
-import { CustomerVehicleHomeComponent } from './page/customer-vehicle-home/customer-vehicle-home.component';
+import { CustomerVehicleHomeComponent } from './pages/customer-vehicle-home/customer-vehicle-home.component';
 
 export const CUSTOMER_VEHICLE_ROUTES: Routes = [
   {
